Convert more erroneous longs to ints

+++ pepper-apis/dsm-core/src/main/java/org/broadinstitute/dsm/db/SmId.java
@@ -33,7 +33,7 @@ public class SmId {
     @ColumnName(DBConstants.SM_ID_TYPE)
     private String smIdType;
     @ColumnName(DBConstants.SM_ID_TISSUE_ID)
-    private Long tissueId;
+    private Integer tissueId;
     @ColumnName(DBConstants.SM_ID_PK)
     private Integer smIdPk;
     @ColumnName(DBConstants.DELETED)
@@ -42,7 +42,7 @@ public class SmId {
     public SmId() {
     }
 
-    public SmId(Integer smIdPk, String smIdType, String smIdValue, Long tissueId) {
+    public SmId(Integer smIdPk, String smIdType, String smIdValue, Integer tissueId) {
         this.smIdPk = smIdPk;
         this.smIdType = smIdType;
         this.smIdValue = smIdValue;
@@ -50,7 +50,7 @@ public class SmId {
     }
 
     public SmId(Integer smIdPk, String smType, String smIdValue, Integer tissueId, Boolean deleted) {
-        this(smIdPk, smType, smIdValue, tissueId.longValue());
+        this(smIdPk, smType, smIdValue, tissueId);
         this.deleted = deleted;
     }
 
@@ -66,7 +66,7 @@ public class SmId {
                     rs.getInt(DBConstants.SM_ID_PK),
                     rs.getString(DBConstants.SM_ID_TYPE_ID),
                     rs.getString(DBConstants.SM_ID_VALUE),
-                    rs.getLong("sm." + DBConstants.TISSUE_ID)
+                    rs.getInt("sm." + DBConstants.TISSUE_ID)
             );
             if (tissueSmId != null) {
                 tissueSmId.setDeleted(rs.getBoolean("sm." + DBConstants.DELETED));
